test(stream-management): add tests for queue desync handling
Add two tests for the queue desync fix:

1. Empty queue with server h > 0 (page reload scenario)
   - Verifies no crash occurs when queue is empty
   - Verifies outbound counter resyncs to server's value

2. Partial queue with server h higher than queue length
   - Verifies existing items are properly acked
   - Verifies counter resyncs after queue is exhausted

Also moves tests to test/ directory to match jest testMatch pattern.
